@@ -1,6 +1,6 @@
 {
-  "name": "@unicef/material-ui-currency-textfield",
-  "version": "0.8.6",
+  "name": "@lupus-ai/mui-currency-textfield",
+  "version": "1.0.0",
   "description": "Currency input textfield for react with Material-ui style",
   "keywords": [
     "material design",
@@ -11,70 +11,45 @@
     "react"
   ],
   "main": "./dist/index.js",
-  "typings": "./dist/index.d.ts",
   "homepage": "https://unicef.github.io/material-ui-currency-textfield",
   "scripts": {
-    "build": "rollup -c",
+    "build": "webpack",
     "start": "rollup -c -w",
     "styleguide": "styleguidist server",
     "styleguide:build": "styleguidist build",
     "deploy": "gh-pages -d styleguide",
     "transpile": "./node_modules/@babel/cli/bin/babel.js src/components/CurrencyTextField -d dist --copy-files",
     "prepublishOnly": "npm run transpile",
-    "publish-demo": "npm run build && npm run deploy"
+    "publish-demo": "npm run build && npm run deploy",
+    "format": "prettier --write \"./src/**/*.{js,jsx}\""
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/unicef/material-ui-currency-textfield.git"
+    "url": "git+https://github.com/LupusAI/mui-currency-textfield.git"
   },
-  "author": "UNICEF",
+  "author": "LupusAI",
   "license": "MIT",
-  "peerDependencies": {
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "prop-types": "^15.7.2",
-    "@material-ui/core": ">= 4.3.0"
-  },
   "devDependencies": {
-    "@babel/cli": "^7.5.5",
-    "@babel/core": "^7.5.5",
-    "@babel/plugin-proposal-object-rest-spread": "^7.5.5",
-    "@babel/plugin-transform-react-jsx": "^7.3.0",
-    "@babel/preset-env": "^7.5.5",
-    "@babel/preset-react": "^7.0.0",
-    "@material-ui/core": "^4.8.3",
-    "@types/node": "^12.6.8",
-    "@types/react": "^16.9.22",
-    "babel-core": "^6.23.0",
-    "babel-loader": "^8.0.5",
-    "babel-plugin-external-helpers": "^6.22.0",
-    "babel-plugin-transform-react-jsx": "^6.24.1",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "babelrc-rollup": "^3.0.0",
-    "core-util-is": "^1.0.2",
-    "file-loader": "^4.1.0",
-    "gh-pages": "^2.1.1",
-    "node-sass": "^4.12.0",
-    "prop-types": "^15.7.2",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-router-dom": "^5.0.1",
-    "react-select": "^3.0.4",
-    "react-styleguidist": "^10.4.2",
-    "rollup": "^1.17.0",
-    "rollup-plugin-babel": "^3.0.7",
-    "rollup-plugin-commonjs": "^10.0.1",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-peer-deps-external": "^2.2.0",
-    "rollup-plugin-url": "^2.2.2",
-    "style-loader": "^0.23.1",
-    "webpack": "^4.29.6",
-    "webpack-cli": "^3.3.6",
-    "webpack-dev-server": "^3.2.1"
+    "@babel/cli": "^7.20.7",
+    "@babel/core": "^7.20.7",
+    "@babel/preset-env": "^7.20.2",
+    "@babel/preset-react": "^7.18.6",
+    "babel-loader": "^9.1.0",
+    "css-loader": "^6.7.3",
+    "gh-pages": "^4.0.0",
+    "prettier": "^2.8.1",
+    "react-styleguidist": "^13.0.0",
+    "style-loader": "^3.3.1",
+    "webpack": "^5.75.0",
+    "webpack-cli": "^5.0.1"
   },
   "dependencies": {
-    "autonumeric": "^4.5.4"
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@mui/material": "^5.11.1",
+    "autonumeric": "^4.6.0",
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "tss-react": "^4.5.0"
   }
 }
