@@ -1,8 +1,9 @@
 const path = require('path')
 const { styles, theme } = require('./styleguide.styles')
+
 module.exports = {
   title: "Material ui currency input",
-  showUsage: true,
+  usageMode: "expand",
   styles,
   theme,
   showSidebar: false,
@@ -34,7 +35,7 @@ module.exports = {
     {
       name: '',
       components: () => ([
-        path.resolve(__dirname, 'src/components/CurrencyTextField', 'CurrencyTextField.js'),
+        path.resolve(__dirname, 'src/components/CurrencyTextField', 'CurrencyTextField.jsx'),
       ])
     },
   ],
